Drafted Abstract & Introduction AND Set up our filtered working dataset
I've completed a basic abstract and adapted the introduction I had for project 1 to fit our research question (and I believe I did an good job with the justification here but I could use some feedback)

Additionally, in the methods section I made a new dataframe which is filtered down to those top 5 neighborhoods and only normal sale conditions and houses built after 1950.
I added some text above this to justify our filtering, but this still needs to be developed a bit.